requirements.txt
tqdm==4.48.2
numpy==1.19.2
scipy==1.6.1
pandas==1.2.2
sklearn==0.23.2
mpmath==1.2.1
bs4==4.9.3
torch-geometric==1.6.1
matplotlib==3.3.1
torch-sparse==0.6.7
torch-scatter==2.0.5
transformers==3.3.1
sentence-transformers==0.3.8
